backend/requirements.txt
flask
flask_cors
bcrypt
PyJWT
psycopg2-binary
python-decouple
python-dotenv